+++ LayerKit.framework/Headers/LYRErrors.h
@@ -29,6 +29,7 @@ typedef NS_ENUM(NSUInteger, LYRError) {
     LYRErrorDistinctConversationExists              = 1013,
     LYRErrorParticipantNotAParticipantInConversation= 1014,
     LYRErrorDistinctDeletedConversationExists       = 1015,
+    LYRErrorParticipantsContainsBlockedUser         = 1016,
     
     // Validation Errors
     LYRErrorInvalidKey                              = 2000,
@@ -64,10 +65,13 @@ typedef NS_ENUM(NSUInteger, LYRClientError) {
     LYRClientErrorDeviceTokenInvalid                = 8000,
     
     // Synchronization Errors
-    LYRClientErrorUndefinedSyncFailure              = 9000,
-    LYRClientErrorDevicePersistenceFailure          = 9001,
-    LYRClientErrorSynchronizationFailure            = 9002,
-    LYRClientErrorManualSyncIgnoredInForeground     = 9003,
+    LYRClientErrorUndefinedSyncFailure                  = 9000,
+    LYRClientErrorDevicePersistenceFailure              = 9001,
+    LYRClientErrorSynchronizationFailure                = 9002,
+    LYRClientErrorManualSyncIgnoredInForeground         = 9003,
+    LYRClientErrorManualSyncFailedNoConnection          = 9004,
+    LYRClientErrorManualSyncIgnoredAlreadyInProgress    = 9005,
+    LYRClientErrorManualSyncIgnoredAlreadyFullySynced   = 9006,
     
     // Debug Errors
     LYRClientErrorZipArchiveCreationFailure         = 10001,
@@ -90,3 +94,9 @@ extern NSString * _Nonnull const LYRErrorUnderlyingErrorsKey;
  @abstract A key into the `userInfo` dictionary of an error returned when attempting to create a new distinct conversation. That key's value represents an existing distinct conversation object.
  */
 extern NSString * _Nonnull const LYRExistingDistinctConversationKey;
+
+/**
+ @abstract A key into the `userInfo` dictionary of an error passed when attempting to run a partial sync for a specific conversation. The key's value represents an `LYRProgress` instance of an existing synchronization process.
+ */
+extern NSString * _Nonnull const LYRExistingSynchronizationProgress;
+